+++ extensions/mongodb/CHANGELOG.md
@@ -7,6 +7,7 @@ Yii Framework 2 mongodb extension Change Log
 - Bug #3385: Fixed "The 'connected' property is deprecated" (samdark)
 - Enh #3520: Added `unlinkAll()`-method to active record to remove all records of a model relation (NmDimas, samdark, cebe)
 - Enh #3778: Gii generator for Active Record model added (klimov-paul)
+- Enh: Migration support added (klimov-paul)
 
 
 2.0.0-beta April 13, 2014